+++ web-client/src/components/Modals/OneTimeModal.tsx
@@ -2,10 +2,11 @@ import { MailOutlined } from '@ant-design/icons';
 import { List, Modal, Typography } from 'antd';
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import r4hLOVE from '../../assets/r4hLOVE.png';
 import { StepForwardButton } from 'src/components/Buttons';
 import { COLORS } from 'src/theme/colors';
 import styled from 'styled-components';
+
+import r4hLOVE from '../../assets/r4hLOVE.png';
 
 const SEEN = 'seen';
