+++ web-client/src/modules/phone/pages/ContentPage.tsx
@@ -3,12 +3,12 @@ import React, { lazy, ReactElement, Suspense, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Redirect, Route, Switch, useLocation } from 'react-router-dom';
 import { observeUserAction } from 'src/ducks/auth/actions';
-import { LoginLocation } from 'src/modules/login/pages/routes/LoginRoute/constants';
 import { PersonalDataLocation } from 'src/modules/personalData/pages/routes/PersonalDataRoute/constants';
 import NotFoundRoute from 'src/pages/routes/NotFoundRoute';
 import { AppState } from 'src/store';
 
 import LoadingWrapper from '../../../components/LoadingComponent/LoadingComponent';
+import { LoginLocation } from '../../registration/pages/routes/LoginRoute/constants';
 import { PhoneEntryLocation } from './routes/PhoneEntryRoute/constants';
 import { PhoneVerifyLocation } from './routes/PhoneVerifyRoute/constants';
