+++ web-client/src/modules/registration/components/GoogleLoginButton/GoogleLoginButton.tsx
