+++ web-client/src/modules/registration/components/RegistrationFooter/RegistrationFooter.tsx
