+++ web-client/src/modules/registration/components/RegistrationSteps/RegistrationSteps.tsx
