+++ web-client/src/modules/registration/components/RegistrationSteps/Step0_Welcome.tsx
