+++ web-client/src/modules/registration/components/RegistrationSteps/Step2_Explanation.tsx
