+++ web-client/src/modules/registration/components/RegistrationSteps/Step3_Authentication.tsx
