+++ web-client/src/modules/registration/pages/ContentPage.tsx
