+++ web-client/src/modules/registration/pages/routes/LoginRoute/LoginRoute.tsx
@@ -3,9 +3,9 @@ import React, { ReactElement } from 'react';
 import { useLocation } from 'react-router-dom';
 import CenteredCard from 'src/components/CenteredCard/CenteredCard';
 import GradientBackground from 'src/components/GradientBackground/GradientBackground';
-import LoginContainer from 'src/modules/login/containers/LoginContainer/LoginContainer';
 
 import IntroWrapper from '../../../../../components/IntroComponent/IntroComponent';
+import LoginContainer from '../../../containers/LoginContainer/LoginContainer';
 
 const LoginRoute: React.FC = (): ReactElement => {
   const location = useLocation();
