+++ web-client/src/pages/routes/ProtectedRoute.tsx
@@ -4,13 +4,13 @@ import { Redirect, Route, RouteProps, useLocation } from 'react-router-dom';
 import { observeUserAction } from 'src/ducks/auth/actions';
 import { observePrivileged, observeProfile } from 'src/ducks/profile/actions';
 import { ProfileState } from 'src/ducks/profile/types';
-import { LoginLocation } from 'src/modules/login/pages/routes/LoginRoute/constants';
 import { PersonalDataLocation } from 'src/modules/personalData/pages/routes/PersonalDataRoute/constants';
 import { RoleInfoLocation } from 'src/modules/personalData/pages/routes/RoleInfoRoute/constants';
 import { PhoneEntryLocation } from 'src/modules/phone/pages/routes/PhoneEntryRoute/constants';
 import { AppState } from 'src/store';
 
 import LoadingWrapper from '../../components/LoadingComponent/LoadingComponent';
+import { LoginLocation } from '../../modules/registration/pages/routes/LoginRoute/constants';
 
 const ProtectedRoute: React.FC<RouteProps> = ({ path, component }) => {
   const auth = useSelector((state: AppState) => state.auth);
